content/html/content/test/test_bug345822.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=345822
-->
<head>
  <title>Test for Bug 345822</title>
  <script type="application/javascript" src="/MochiKit/packed.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <style>
    input, textarea { background-color: rgb(0,0,0) !important; }
    :valid   { background-color: rgb(0,255,0) !important; }
    :invalid { background-color: rgb(255,0,0) !important; }
  </style>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=345822">Mozilla Bug 345822</a>
<p id="display"></p>
<div id="content">
  <form>
  </form>
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 345822 **/

function checkRequiredAttribute(element)
{
  ok('required' in element, "Element should have the required attribute");

  ok(!element.required, "Element required attribute should be disabled");
  is(element.getAttribute('required'), null,
    "Element required attribute should be disabled");

  element.required = true;
  ok(element.required, "Element required attribute should be enabled");
  isnot(element.getAttribute('required'), null,
    "Element required attribute should be enabled");

  element.removeAttribute('required');
  element.setAttribute('required', '');
  ok(element.required, "Element required attribute should be enabled");
  isnot(element.getAttribute('required'), null,
    "Element required attribute should be enabled");

  element.removeAttribute('required');
  ok(!element.required, "Element required attribute should be disabled");
  is(element.getAttribute('required'), null,
    "Element required attribute should be disabled");
}

function checkNotSufferingFromBeingMissing(element, doNotApply)
{
  ok(!element.validity.valueMissing,
    "Element should not suffer from value missing");
  ok(element.validity.valid, "Element should be valid");
  ok(element.checkValidity(), "Element should be valid");
  is(element.validationMessage, "",
    "Validation message should be the empty string");

  if (element.type != 'radio' && element.type != 'checkbox') {
    is(window.getComputedStyle(element, null).getPropertyValue('background-color'),
       doNotApply ? "rgb(0, 0, 0)" : "rgb(0, 255, 0)",
       "The pseudo-class is not correctly applied");
  }
}

function checkSufferingFromBeingMissing(element)
{
  ok(element.validity.valueMissing, "Element should suffer from value missing");
  ok(!element.validity.valid, "Element should not be valid");
  ok(!element.checkValidity(), "Element should not be valid");

  if (element.type == 'checkbox')
  {
    is(element.validationMessage,
      "Please check this box if you want to proceed.",
      "Validation message is wrong");
  }
  else if (element.type == 'radio')
  {
    is(element.validationMessage,
      "Please select one of these options.",
      "Validation message is wrong");
  }
  else if (element.type == 'file')
  {
    is(element.validationMessage,
      "Please select a file.",
      "Validation message is wrong");
  }
  else // text fields
  {
    is(element.validationMessage,
      "Please fill out this field.",
      "Validation message is wrong");
  }

  if (element.type != 'radio' && element.type != 'checkbox') {
    is(window.getComputedStyle(element, null).getPropertyValue('background-color'),
       "rgb(255, 0, 0)", ":invalid pseudo-class should apply");
  }
}

function checkTextareaRequiredValidity()
{
  var element = document.createElement('textarea');
  document.forms[0].appendChild(element);

  element.value = '';
  element.required = false;
  checkNotSufferingFromBeingMissing(element);

  element.required = true;
  checkSufferingFromBeingMissing(element);

  element.readOnly = true;
  checkNotSufferingFromBeingMissing(element, true);

  element.readOnly = false;
  checkSufferingFromBeingMissing(element);

  element.value = 'foo';
  checkNotSufferingFromBeingMissing(element);

  element.value = '';
  checkSufferingFromBeingMissing(element);

  element.required = false;
  checkNotSufferingFromBeingMissing(element);

  document.forms[0].removeChild(element);
}

function checkInputRequiredNotApply(type, isBarred)
{
  var element = document.createElement('input');
  element.type = type;
  document.forms[0].appendChild(element);

  element.value = '';
  element.required = false;
  checkNotSufferingFromBeingMissing(element, isBarred);

  element.required = true;
  checkNotSufferingFromBeingMissing(element, isBarred);

  element.required = false;

  document.forms[0].removeChild(element);
}

function checkInputRequiredValidity(type)
{
  var element = document.createElement('input');
  element.type = type;
  document.forms[0].appendChild(element);

  element.value = '';
  element.required = false;
  checkNotSufferingFromBeingMissing(element);

  element.required = true;
  checkSufferingFromBeingMissing(element);

  element.readOnly = true;
  checkNotSufferingFromBeingMissing(element, true);

  element.readOnly = false;
  checkSufferingFromBeingMissing(element);

  if (element.type == 'email') {
    element.value = 'foo@bar.com';
  } else if (element.type == 'url') {
    element.value = 'http://mozilla.org/';
  } else {
    element.value = 'foo';
  }
  checkNotSufferingFromBeingMissing(element);

  element.value = '';
  checkSufferingFromBeingMissing(element);

  document.forms[0].removeChild(element);
  checkNotSufferingFromBeingMissing(element);
}

function checkInputRequiredValidityForCheckbox()
{
  var element = document.createElement('input');
  element.type = 'checkbox';
  document.forms[0].appendChild(element);

  element.checked = false;
  element.required = false;
  checkNotSufferingFromBeingMissing(element);

  element.required = true;
  checkSufferingFromBeingMissing(element);

  element.checked = true;
  checkNotSufferingFromBeingMissing(element);

  element.checked = false;
  checkSufferingFromBeingMissing(element);

  element.required = false;
  checkNotSufferingFromBeingMissing(element);

  document.forms[0].removeChild(element);
}

function checkInputRequiredValidityForRadio()
{
  var element = document.createElement('input');
  element.type = 'radio';
  element.name = 'test'
  document.forms[0].appendChild(element);

  element.checked = false;
  element.required = false;
  checkNotSufferingFromBeingMissing(element);

  element.required = true;
  checkSufferingFromBeingMissing(element);

  element.checked = true;
  checkNotSufferingFromBeingMissing(element);

  element.checked = false;
  checkSufferingFromBeingMissing(element);

  // A required radio button should not suffer from value missing if another
  // radio button from the same group is checked.
  var element2 = document.createElement('input');
  element2.type = 'radio';
  element2.name = 'test';

  element2.checked = true;
  element2.required = false;
  document.forms[0].appendChild(element2);

  // Adding a checked radio should make required radio in the group not
  // suffering from being missing.
  checkNotSufferingFromBeingMissing(element);

  element.checked = false;
  element2.checked = false;
  checkSufferingFromBeingMissing(element);

  // The other radio button should not be disabled.
  // A disabled checked radio button in the radio group
  // is enough to not suffer from value missing.
  element2.checked = true;
  element2.disabled = true;
  checkNotSufferingFromBeingMissing(element);

  // If a radio button is not required but another radio button is required in
  // the same group, the not required radio button should not suffer from value
  // missing.
  element2.disabled = false;
  element2.checked = false;
  element.required = false;
  element2.required = true;
  checkNotSufferingFromBeingMissing(element);
  checkSufferingFromBeingMissing(element2);

  element.checked = true;
  checkNotSufferingFromBeingMissing(element2);

  // The checked radio is not in the group anymore, element2 should be invalid.
  document.forms[0].removeChild(element);
  checkSufferingFromBeingMissing(element2);

  document.forms[0].removeChild(element2);
}

function checkInputRequiredValidityForFile()
{
  var element = document.createElement('input');
  element.type = 'file'
  document.forms[0].appendChild(element);

  function createFileWithData(fileName, fileData) {
    var dirSvc = Components.classes["@mozilla.org/file/directory_service;1"]
                           .getService(Components.interfaces.nsIProperties);
    var testFile = dirSvc.get("ProfD", Components.interfaces.nsIFile);
    testFile.append(fileName);
    var outStream = Components.
                    classes["@mozilla.org/network/file-output-stream;1"].
                    createInstance(Components.interfaces.nsIFileOutputStream);
    outStream.init(testFile, 0x02 | 0x08 | 0x20, // write, create, truncate
                   0666, 0);
    outStream.write(fileData, fileData.length);
    outStream.close();

    return testFile;
  }

  // Need privileges to set a filename with .value.
  netscape.security.PrivilegeManager.enablePrivilege("UniversalXPConnect");

  var file = createFileWithData("345822_file", "file content");

  element.value = "";
  element.required = false;
  checkNotSufferingFromBeingMissing(element);

  element.required = true;
  checkSufferingFromBeingMissing(element);

  element.value = file.path;
  checkNotSufferingFromBeingMissing(element);

  element.value = "";
  checkSufferingFromBeingMissing(element);

  element.required = false;
  checkNotSufferingFromBeingMissing(element);

  file.remove(false);
  document.forms[0].removeChild(element);
}

checkRequiredAttribute(document.createElement('textarea'));
checkTextareaRequiredValidity();

// Every input element should have the required attribute.
checkRequiredAttribute(document.createElement('input'));

// The require attribute behavior depend of the input type.
// First of all, checks for types that make the element barred from
// constraint validation.
var typeBarredFromConstraintValidation = ["hidden", "button", "reset"];
for each (type in typeRequireNotApply) {
  checkInputRequiredNotApply(type, true);
}

// Then, checks for the types which do not use the required attribute.
// TODO: check 'color' and 'range' when they will be implemented.
var typeRequireNotApply = ["submit", "image"];
for each (type in typeRequireNotApply) {
  checkInputRequiredNotApply(type, false);
}

// Now, checking for all types which accept the required attribute.
// TODO: check 'datetime', 'date', 'month', 'week', 'time', 'datetime-local'
//       and 'number' when they will be implemented.
var typeRequireApply = ["text", "password", "search", "tel", "email", "url"];

for each (type in typeRequireApply) {
  checkInputRequiredValidity(type);
}

checkInputRequiredValidityForCheckbox();
checkInputRequiredValidityForRadio();
checkInputRequiredValidityForFile();

</script>
</pre>
</body>
</html>
